Separates metric exceptions.
